refactor: move statistics functions to math module
