+++ index.ts
@@ -1,12 +1,5 @@
 export * from './arrayFunctions/arrayDifference';
 export * from './arrayFunctions/arrayIntersection';
-export * from './arrayFunctions/calculateAverage';
-export * from './arrayFunctions/calculateMedian';
-export * from './arrayFunctions/calculateMode';
-export * from './arrayFunctions/calculateProduct';
-export * from './arrayFunctions/calculateRange';
-export * from './arrayFunctions/calculateStandardDeviation';
-export * from './arrayFunctions/calculateVariance';
 export * from './arrayFunctions/cartesianProduct';
 export * from './arrayFunctions/chunkArray';
 export * from './arrayFunctions/findCommonWithCondition';
@@ -93,6 +86,13 @@ export * from './mathFunctions/lcm';
 export * from './mathFunctions/oddOrEven';
 export * from './mathFunctions/roundToDecimals';
 export * from './mathFunctions/roundValue';
+export * from './mathFunctions/statisticsFunctions/calculateAverage';
+export * from './mathFunctions/statisticsFunctions/calculateMedian';
+export * from './mathFunctions/statisticsFunctions/calculateMode';
+export * from './mathFunctions/statisticsFunctions/calculateProduct';
+export * from './mathFunctions/statisticsFunctions/calculateRange';
+export * from './mathFunctions/statisticsFunctions/calculateStandardDeviation';
+export * from './mathFunctions/statisticsFunctions/calculateVariance';
 export * from './mathFunctions/geometricFunctions/calculateCircleArea';
 export * from './mathFunctions/geometricFunctions/calculateCircleCircumference';
 export * from './mathFunctions/geometricFunctions/calculateConeSurfaceArea';
