+++ mathFunctions/statisticsFunctions/calculateMode.ts
