+++ mathFunctions/statisticsFunctions/calculateProduct.ts
