+++ mathFunctions/statisticsFunctions/calculateRange.ts
