+++ mathFunctions/statisticsFunctions/calculateStandardDeviation.ts
