+++ mathFunctions/statisticsFunctions/calculateVariance.ts
